# btf: introduce Bits type and remove Int.Bits and Int.Offset