--- a/internal/btf/strings_test.go
+++ b/internal/btf/strings_test.go
@@ -59,3 +59,15 @@ func TestStringTable(t *testing.T) {
 		t.Fatal("Accepted non-empty first item")
 	}
 }
+
+func newStringTable(strings ...string) *stringTable {
+	offsets := make([]uint32, len(strings))
+
+	var offset uint32
+	for i, str := range strings {
+		offsets[i] = offset
+		offset += uint32(len(str)) + 1 // account for NUL
+	}
+
+	return &stringTable{offsets, strings}
+}