@lmb 2022-05-10
Added `TestInflateLegacyBitfield` after discussing with Timo.